+++ release-notes.md
@@ -3,4 +3,5 @@
 ## 1.5.XX
 - BREAKING CHANGE: currencyFormat renamed to negativeStyle
 - BREAKING CHANGE: negativeStyle type "accounting" changed to "parentheses"
+- CurrencyInput does not support instant mode anymore
 - Added different negative styles for ingestion ("default" | "parentheses" | "redParentheses" | "red")